"New Chapter" Codes in Text Format(GML)
The movement function works by firstly creating variables of "W, A, S, and D" input. Since GML works with -1, 0, and 1 in input, what I did, is first I was checking in which direction the player is going to move in, that is what the variables "moveh" and "movev" are for, after which we multiply these variables with the designated speed. However, so far we have only declared the speed, the actual movement is called after both horizontal and vertical collision checks. The reason we need a specific speed for diagonal movement is that Game Maker adds both speeds on top of each other, so we had to make a specific diagonal speed variable with some tweaks from the original. The way the collision check works is it looks for the distance between the player and the collider, which in this case is "o_wall", before coming to a stop.